pyproject.toml
[tool.poetry]
name = "e2e"
version = "0.1.0"
description = "E2E tests for AIOps"
authors = ["Harish Balakrishnan <harishbalakrishnan3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
PyHamcrest = "^2.1.0"
PyYaml = "^6.0.2"
behave = "^1.2.6"
mockseries = "^0.3.1"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"
opentelemetry-instrumentation = "^0.49b2"
opentelemetry-instrumentation-openai = "^0.33.1"
opentelemetry-exporter-prometheus-remote-write = "^0.49b2"
opentelemetry-semantic-conventions = "^0.49b2"
opentelemetry-semantic-conventions-ai = "^0.4.1"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
regex = "^2024.9.11"
python-dateutil = "^2.9.0.post0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
